# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Catch API
  version: 0.1.0
  description: Backend contracts for Catch modular monolith.
servers:
  - url: /api/v1
...
paths:
...
  /moderation/submissions:
    get:
      tags: [Moderation]
      summary: List pending article moderation submissions
      operationId: listModerationSubmissions
      security:
        - cookieSession: []
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Pending submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationSubmissionList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /moderation/submissions/{submissionID}/approve:
    post:
      tags: [Moderation]
      summary: Approve article moderation submission
      operationId: approveModerationSubmission
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - name: submissionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Submission state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationSubmission"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /moderation/submissions/{submissionID}/reject:
    post:
      tags: [Moderation]
      summary: Reject article moderation submission
      operationId: rejectModerationSubmission
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - name: submissionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectModerationSubmissionRequest"
      responses:
        "200":
          description: Submission state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationSubmission"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /moderation/submissions/{submissionID}/threads:
    get:
      tags: [Moderation]
      summary: List moderation threads for submission
      operationId: listModerationThreads
      security:
        - cookieSession: []
      parameters:
        - name: submissionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Moderation threads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationThreadList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Moderation]
      summary: Create moderation thread
      operationId: createModerationThread
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - name: submissionID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModerationThreadRequest"
      responses:
        "201":
          description: Created moderation thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationThread"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /moderation/threads/{threadID}/resolve:
    post:
      tags: [Moderation]
      summary: Resolve moderation thread
      operationId: resolveModerationThread
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - name: threadID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Thread state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationThread"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /moderation/threads/{threadID}/reopen:
    post:
      tags: [Moderation]
      summary: Reopen moderation thread
      operationId: reopenModerationThread
      security:
        - cookieSession: []
        - csrfHeader: []
      parameters:
        - name: threadID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReopenModerationThreadRequest"
      responses:
        "200":
          description: Thread state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationThread"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: catch_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
...
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated user has no permission for this action
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationFailed:
      description: Validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    RejectModerationSubmissionRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 2000
    CreateModerationThreadRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        block_id:
          type: string
          maxLength: 128
        body:
          type: string
          minLength: 1
          maxLength: 4000
    ReopenModerationThreadRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          maxLength: 2000
    ModerationSubmission:
      type: object
      additionalProperties: false
      required: [id, article_id, revision_id, author_id, status, rejection_reason, approval_count, open_thread_count, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        article_id:
          type: string
          format: uuid
        revision_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        rejection_reason:
          type: string
        approval_count:
          type: integer
          minimum: 0
        open_thread_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ModerationSubmissionList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModerationSubmission"
    ModerationThread:
      type: object
      additionalProperties: false
      required: [id, submission_id, author_id, body, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        submission_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        block_id:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [open, resolved]
        created_at:
          type: string
          format: date-time
    ModerationThreadList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModerationThread"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [type, title, status, code, message]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          enum:
            - invalid_request
            - validation_failed
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - not_implemented
            - service_unavailable
            - internal_error
        message:
          type: string
        trace_id:
          type: string
        details:
          type: object
          additionalProperties: true